Adding confirm before action

+++ CMS/admin/includes/view_all_posts.php
@@ -1,12 +1,10 @@
 <?php
 
 if (isset($_POST['checkBoxArray'])) {
+    foreach ($_POST['checkBoxArray'] as $checkBoxValue) {
+        $bulk_options = $_POST['bulk_options'];
 
-  foreach ($_POST['checkBoxArray'] as $checkBoxValue) {
-
-   $bulk_options = $_POST['bulk_options'];
-
-   switch ($bulk_options) {
+        switch ($bulk_options) {
 
      case 'published':
 
@@ -35,7 +33,7 @@ if (isset($_POST['checkBoxArray'])) {
        # code...
        break;
    }
-  }
+    }
 }
 
  ?>
@@ -86,20 +84,17 @@ if (isset($_POST['checkBoxArray'])) {
     $select_posts = mysqli_query($connection, $query);
 
       while ($row = mysqli_fetch_assoc($select_posts)) {
+          $post_id = $row['post_id'];
+          $post_author = $row['post_author'];
+          $post_title = $row['post_title'];
+          $post_category_id = $row['post_category_id'];
+          $post_status = $row['post_status'];
+          $post_image = $row['post_image'];
+          $post_tags = $row['post_tags'];
+          $post_comment_count = $row['post_comment_count'];
+          $post_date = $row['post_date'];
 
-        $post_id = $row['post_id'];
-        $post_author = $row['post_author'];
-        $post_title = $row['post_title'];
-        $post_category_id = $row['post_category_id'];
-        $post_status = $row['post_status'];
-        $post_image = $row['post_image'];
-        $post_tags = $row['post_tags'];
-        $post_comment_count = $row['post_comment_count'];
-        $post_date = $row['post_date'];
-
-        echo "<tr>";
-
-        ?>
+          echo "<tr>"; ?>
 
         <td><input class='checkBoxes' type='checkbox' name='checkBoxArray[]' value='<?php echo $post_id; ?>'></td>
 
@@ -107,29 +102,28 @@ if (isset($_POST['checkBoxArray'])) {
 
 
         echo "<td>$post_id</td>";
-        echo "<td>$post_author</td>";
-        echo "<td>$post_title</td>";
+          echo "<td>$post_author</td>";
+          echo "<td>$post_title</td>";
 
-        $query = "SELECT * FROM categories WHERE cat_id = $post_category_id ";
-        $select_categories_id = mysqli_query($connection, $query);
+          $query = "SELECT * FROM categories WHERE cat_id = $post_category_id ";
+          $select_categories_id = mysqli_query($connection, $query);
 
-        while ($row = mysqli_fetch_assoc($select_categories_id)) {
-        $cat_id = $row['cat_id'];
-        $cat_title = $row['cat_title'];
+          while ($row = mysqli_fetch_assoc($select_categories_id)) {
+              $cat_id = $row['cat_id'];
+              $cat_title = $row['cat_title'];
 
-        echo "<td>$cat_title</td>";
-}
+              echo "<td>$cat_title</td>";
+          }
 
-        echo "<td>$post_status</td>";
-        echo "<td><img width='85' class='img-responsive' src='../images/$post_image' alt='image'></td>";
-        echo "<td>$post_tags</td>";
-        echo "<td>$post_comment_count</td>";
-        echo "<td>$post_date</td>";
-        echo "<td><a href='../post.php?&p_id={$post_id}'>View Post</a></td>";
-        echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-        echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
-        echo "<tr>";
-
+          echo "<td>$post_status</td>";
+          echo "<td><img width='85' class='img-responsive' src='../images/$post_image' alt='image'></td>";
+          echo "<td>$post_tags</td>";
+          echo "<td>$post_comment_count</td>";
+          echo "<td>$post_date</td>";
+          echo "<td><a href='../post.php?&p_id={$post_id}'>View Post</a></td>";
+          echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
+          echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete this post ?');\" href='posts.php?delete={$post_id}'>Delete</a></td>";
+          echo "<tr>";
       }
 
      ?>
@@ -140,13 +134,11 @@ if (isset($_POST['checkBoxArray'])) {
 <?php
 
 if (isset($_GET['delete'])) {
+    $the_post_id = $_GET['delete'];
 
-   $the_post_id = $_GET['delete'];
-
-   $query = "DELETE FROM posts WHERE post_id = {$the_post_id}";
-   $delete_query = mysqli_query($connection, $query);
-   header("Location: posts.php");
-
+    $query = "DELETE FROM posts WHERE post_id = {$the_post_id}";
+    $delete_query = mysqli_query($connection, $query);
+    header("Location: posts.php");
 }
